Remove duplicate references with vfbd_parser
Included a list-based filter for duplicate sequence ids in the downloaded VFDB fasta file. As consequence, ´ariba prepareref´ can be run after execution of vfdb_parser without manual deletion of duplicate entries.
Validated with VFDB_setB_nt.fas.gz downloaded on 2022-01-18. Command ´ariba prepareref´ is not raising error because of duplicate seq.id (but 1254 sequences are filtered out because they are not recognized as genes though).